clean up etherscan verification logic
